Add an interface for providing custom approval logic.
