+++ cmd/identityoracled/handlers/transfers.go
@@ -5,6 +5,7 @@ import (
 	"net/http"
 
 	"github.com/tokenized/identity-oracle/internal/oracle"
+	"github.com/tokenized/identity-oracle/internal/platform/config"
 	"github.com/tokenized/identity-oracle/internal/platform/db"
 	"github.com/tokenized/identity-oracle/internal/platform/web"
 
@@ -21,6 +22,7 @@ type Transfers struct {
 	MasterDB     *db.DB
 	Key          bitcoin.Key
 	BlockHandler *oracle.BlockHandler
+	Approver     oracle.ApproverInterface
 }
 
 // TransferSignature returns an approve/deny signature for a transfer receiver.
@@ -45,10 +47,37 @@ func (t *Transfers) TransferSignature(ctx context.Context, log logger.Logger, w 
 	dbConn := t.MasterDB.Copy()
 	defer dbConn.Close()
 
+	user, err := oracle.FetchUserByXPub(ctx, dbConn, requestData.XPubs)
+	if err != nil {
+		if errors.Cause(err) == oracle.ErrXPubNotFound {
+			web.RespondError(ctx, log, w, err, http.StatusNotFound)
+			return nil
+		}
+		return translate(errors.Wrap(err, "fetch user"))
+	}
+
+	var approved bool
+	var description string
+	if t.Approver != nil {
+		var approveErr error
+		approved, description, approveErr = t.Approver.ApproveTransfer(ctx, requestData.Contract,
+			requestData.AssetID, requestData.Quantity, user.ID)
+		if approveErr != nil {
+			return translate(errors.Wrap(err, "approver"))
+		}
+	}
+
+	// Dev reject testing
+	testValues := config.ContextTestValues(ctx)
+	if testValues.RejectQuantity != 0 && testValues.RejectQuantity == requestData.Quantity {
+		approved = false
+		description = "test reject quantity"
+	}
+
 	// Check that xpub is in DB. Check that entity associated xpub meets criteria for asset.
-	sigHash, height, hash, approved, err := oracle.ApproveTransfer(ctx, dbConn, t.BlockHandler,
+	sigHash, height, hash, err := oracle.CreateReceiveSignature(ctx, dbConn, t.BlockHandler,
 		requestData.Contract, requestData.AssetID, requestData.XPubs, requestData.Index,
-		requestData.Quantity)
+		requestData.Quantity, approved)
 	if err != nil {
 		return translate(errors.Wrap(err, "approve transfer"))
 	}
@@ -60,12 +89,14 @@ func (t *Transfers) TransferSignature(ctx context.Context, log logger.Logger, w 
 
 	response := struct {
 		Approved     bool           `json:"approved"`
+		Description  string         `json:"description"`
 		SigAlgorithm uint32         `json:"algorithm"`
 		Sig          string         `json:"signature"`
 		BlockHeight  uint32         `json:"block_height"`
 		BlockHash    bitcoin.Hash32 `json:"block_hash"`
 	}{
 		Approved:     approved,
+		Description:  description,
 		SigAlgorithm: 1,
 		Sig:          sig.String(),
 		BlockHeight:  height,
